suite change (#7822)
